requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
requests>=2.32,<3.0
gspread>=6.0,<7.0
oauth2client>=4.1.3,<5.0
google-api-python-client>=2.170,<3.0
selenium>=4.21,<5.0
webdriver-manager>=4.0.2,<5.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
